Move ArgoCD upgrade
