+++ upgrades/0.0.87.argocd/context.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"github.com/oslokommune/okctl-upgrade/0.0.87.argocd/pkg/lib/cmdflags"
-	"github.com/oslokommune/okctl-upgrade/0.0.87.argocd/pkg/lib/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.87.argocd/pkg/lib/cmdflags"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.87.argocd/pkg/lib/logger"
 )
 
 type Context struct {
