+++ upgrades/0.0.87.argocd/pkg/argocd/api.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"github.com/miekg/dns"
 	merrors "github.com/mishudark/errors"
-	"github.com/oslokommune/okctl-upgrade/0.0.87.argocd/pkg/lib/cmdflags"
-	"github.com/oslokommune/okctl-upgrade/0.0.87.argocd/pkg/lib/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.87.argocd/pkg/lib/cmdflags"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.87.argocd/pkg/lib/logger"
 	"github.com/oslokommune/okctl/pkg/api"
 	"github.com/oslokommune/okctl/pkg/cfn"
 	"github.com/oslokommune/okctl/pkg/client"
