+++ upgrades/0.0.87.argocd/pkg/argocd/cluster_declaration.go
