+++ upgrades/0.0.87.argocd/pkg/lib/commonerrors/api.go
